+++ src/frontend/src/components/home/employees/__tests__/ManageEmployeesT.ts
@@ -1,6 +1,6 @@
 import { createVuetify } from 'vuetify'
 import { mount } from 'cypress/vue'
-import ManageEmployeesT from '../management/ManageEmployeesT.vue'
+import ManageEmployeesT from '../management/ManageEmployees.vue'
 
 it('renders manage employees', () => {
   const vuetify = createVuetify()
